Feature/support web backend api

+++ SupportWeb.Data/Migrations/SupportShopDbContextModelSnapshot.cs
@@ -358,8 +358,8 @@ namespace SupportWeb.Data.Migrations
                         .HasColumnType("nvarchar(200)")
                         .HasMaxLength(200);
 
-                    b.Property<int>("FileSize")
-                        .HasColumnType("int");
+                    b.Property<long>("FileSize")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("ImagePath")
                         .IsRequired()
